requirements.txt
pycryptodomex==3.19.0
pyusb==1.2.1
websockets==12.0
